+++ changelog.md
@@ -43,6 +43,7 @@
 * Added password protection for publicly shared grids and forms.
 * Made views trashable.
 * Fixed bug where a cell value was not reverted when the request to the backend fails.
+* **Premium** Added row coloring.
 
 ## Released (2022-03-03 1.9.1)